📦️ Move domains into apps
