+++ templates/quiz/admin/daily_quiz_list.html
