+++ templates/user/homepage.html
@@ -1,4 +1,4 @@
-{% extends "question/user/base.html" %}
+{% extends "question/user/../question/user/base.html" %}
 {% load static %}
 
 {% block title %}홈페이지 - Tok! CS{% endblock %}
